+++ packages/crypto/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @owf/crypto
 
+## 0.3.2
+
+### Patch Changes
+
+- @owf/identity-common@0.3.2
+
 ## 0.3.1
 
 ### Patch Changes