+++ packages/eudi-lote/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @owf/eudi-lote
 
+## 0.3.2
+
+### Patch Changes
+
+- d87e13b: Simplifies the errors returned by `validateLoTEProfile` by pre-matching the given LoTE to the corresponding profiles' type identifiers:
+
+  - When no type matches, the error clearly states that the LoTE Type doesn't match any of the given profiles.
+  - When one type matches, but fails along the verification, only the errors of this profile are returned for clarity.
+  - @owf/crypto@0.3.2
+  - @owf/identity-common@0.3.2
+
 ## 0.3.1
 
 ### Patch Changes