+++ packages/token-status-list/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @owf/token-status-list
 
+## 0.3.2
+
+### Patch Changes
+
+- @owf/cose@0.3.2
+- @owf/identity-common@0.3.2
+
 ## 0.3.1
 
 ### Patch Changes